-Fixed player/random crashes part #1 -Added Debug few for localplayer and one networkplayer(dev) -Set stream distance of players to -1(no streamout), until XForce has fixed the streamer code -Renamed CCrashFixes into CHooks and added/disabled some code -Removed IVMP_WEBKIT from preprocessor compile stuff -Fixed/Improved loadingscreen(background alpha & position of labels) -Reversed some player tasks/natives(lookat, move etc.) -Rewritten OnFoot sync(part #1) - walking/running should be fluent now -Reversed StartWakling/StopWakling ped-tasks -Cleaned up some code
NOTE: this is a development version, not recommended for playing!

+++ Client/Core/CGUI.cpp
@@ -712,7 +712,7 @@ std::string CGUIWindow::getText()
 
 	char * Ansi = new char[str.length() + 1];
 	std::string out;
-	for(int i = 0; i < str.size(); i++)
+	for(int i = 0; i < (int)str.size(); i++)
 	{
 		Ansi[i] = (UCHAR)str[i];
 	}
